*: delete most of the code
This was really painful to make and possibly wrong. Most of the time
spent is compiling this massive models/ package, so I trimmed it by hand by
doing:
rm -rf models/

then checking code back out until it compiled again.

Unfortunately, at time of writing the kiota downloader is broken..

A different approach to generating this commit might be to instead
edit the downloaded yaml down and generate with kiota to have better
correctness guarantees.